+++ packages/shared/src/constants/index.ts
@@ -1,5 +1,5 @@
 export const APP_NAME = 'NextDestiny BattlePlanner';
-export const APP_VERSION = '1.0.0';
+export const APP_VERSION = '3.0.0';
 
 export const COLORS_ARRAY = [
   '#FF0000', // Red